requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
Werkzeug==0.10.4
itsdangerous==0.24
python-aalib==0.3
wsgiref==0.1.2
